pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "gst-filing-system"
version = "1.0.0"
description = "AI-Powered GST Return Filing System with Document Processing"
authors = [{name = "ADK Team", email = "team@adk.com"}]
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0",
    "python-multipart>=0.0.6",
    "python-dotenv>=1.0.0",
    "google-generativeai>=0.3.2",
    "markitdown>=0.0.1a2",
    "pydantic>=2.5.0",
    "sqlalchemy>=2.0.0",
    "streamlit>=1.28.1",
    "requests>=2.31.0",
    "pandas>=2.0.0",
    "openpyxl>=3.1.0",
    "reportlab>=3.6.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.6",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.0.0",
]

[project.scripts]
start-backend = "uvicorn main:app --reload --host 0.0.0.0 --port 8000"
start-ui = "streamlit run ui/app.py --server.port 8501"

[tool.setuptools.packages.find]
where = ["."]
include = ["routes*", "usecases*", "models*", "schemas*", "agents*", "database*", "ui*"]

[tool.black]
line-length = 88
target-version = ['py38']

[tool.isort]
profile = "black"
